requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
defusedxml==0.7.1
distro==1.9.0
Django==6.0.4
django-allauth==65.16.1
docstring_parser==0.17.0
easyocr==1.7.2
filelock==3.25.2
fsspec==2026.3.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-documentai==3.14.0
google-cloud-resource-manager==1.17.0
google-cloud-storage==2.19.0
google-cloud-vision==3.13.0
google-crc32c==1.8.0
google-genai==1.72.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.4
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vertexai==1.71.1
websockets==16.0
webvtt-py==0.5.1
youtube-transcript-api==1.2.4
yt-dlp==2026.3.17
google-cloud-texttospeech==2.16.2
celery==5.4.0
redis==5.2.1
